giang - code lại các trang jsp và update hoàn thành chức năng quen mật khẩu và thay đổi mật khẩu chỉnh lại trang index, forgot change thêm validate cho trang change pass

+++ src/controller/CompleteChangePass.java
@@ -12,12 +12,16 @@ import java.io.IOException;
 @WebServlet("/CompleteChangePass")
 public class CompleteChangePass extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String email = request.getParameter("email");
-        System.out.println("email là: "+email);
-        String pass = request.getParameter("pass");
-        Database.changePass(email,pass);
-        Database.deleteToken(email);
-        System.out.println(" thay đổi mật khẩu thành công");
+      String email = request.getParameter("email");
+             System.out.println("email là: " + email);
+             if (email != null) {
+                 String pass = request.getParameter("pass");
+                 Database.changePass(email, pass);
+                 Database.deleteToken(email);
+                 ///chuyern trang login
+                 response.sendRedirect("http://localhost:8080/ParkingLot/login/loginForm.jsp");
+                 System.out.println(" thay đổi mật khẩu thành công");
+             }
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
